src/python/workshop/requirements.txt
Faker>=37.4.0, <38.0.0
aiohttp>=3.12.13,<4.0.0
asyncpg>=0.30.0,<0.31.0
azure-ai-agents==1.1.0b4
azure-ai-projects==1.0.0b12
azure-core-tracing-opentelemetry>=1.0.0b12, <2.0.0
azure-identity>=1.23.0,<2.0.0
azure-monitor-opentelemetry>=1.6.10,<2.0.0
fastapi>=0.115.14,<0.116.0
httpx>=0.28.1,<0.29.0
mcp>=1.10.0,<2.0.0
openai>=1.98.0,<2.0.0
opentelemetry-sdk>=1.31.1,<2.0.0
pandas>=2.3.0,<3.0.0
python-dotenv>=1.1.1,<2.0.0
python-multipart>=0.0.20, <0.0.30
uvicorn[standard]>=0.35.0,<0.36.0